Add files via upload
Not sure if website will run, since it was hosted on Harvard CS50's old IDE that's probably phased out by now, but it did run and it was my biggest website project yet.  This was done on the backend with Python, Flask, SQLite, and used an API to get the stock quotes.  Each user can create their own account if they use a unique username with a valid password.  All stocks "purchased" were kept in their individual database, and can be retrieved as just info or to modify due to buying/selling.  I'm 90% sure all the error checks were done both on front end via JavaScript and also on the backend with Python.  I really enjoyed this project, it was fun and appropriately challenging.